Rename betty.content to betty.content_builder
