+++ betty/content_builders/raspberry_mint_columns.py
@@ -15,7 +15,7 @@ from betty.assertions.mapping import assert_mapping
 from betty.assertions.sequence import assert_sequence
 from betty.asset_directories.raspberry_mint import raspberry_mint
 from betty.attrs.owner import OwnerAttr
-from betty.content import (
+from betty.content_builder import (
     ContentBuilder,
     ContentBuilderDefinition,
     ContentBuilderManufacturer,
