+++ betty/content_builders/static.py
@@ -6,7 +6,7 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING, final, override
 
-from betty.content import ContentBuilder, ContentBuilderDefinition
+from betty.content_builder import ContentBuilder, ContentBuilderDefinition
 from betty.locale.localizable.gettext import _
 
 if TYPE_CHECKING:
