+++ betty/jinja_filters/build_content.py
@@ -9,7 +9,7 @@ from typing import TYPE_CHECKING, Self, final, override
 from jinja2 import pass_context
 from markupsafe import Markup
 
-from betty.content import ContentBuilder, build
+from betty.content_builder import ContentBuilder, build
 from betty.factory import Factory, Manufacturable
 from betty.jinja import context_document
 from betty.jinja.filter import JinjaFilter, JinjaFilterDefinition
