+++ betty/tests/content_builders/test_raspberry_mint_color_style.py
@@ -1,4 +1,4 @@
-from betty.content import ContentBuilderManufacturer
+from betty.content_builder import ContentBuilderManufacturer
 from betty.content_builders.raspberry_mint_color_style import (
     ColorStyle,
     ColorStyleData,
