+++ betty/tests/jinja_filters/test_build_content.py
@@ -2,7 +2,7 @@ from collections.abc import Iterable
 
 import pytest
 
-from betty.content import ContentBuilder, ContentBuilderDefinition
+from betty.content_builder import ContentBuilder, ContentBuilderDefinition
 from betty.content_builders.static import Static
 from betty.plugin.factory import PluginManufacturer
 from betty.test_utils.conftest import AssertTemplateString
